Fixes to CSS in featured and nav
For featured: Boostrap grid has been used to align the event headings
and the event card rather than the media queries, which however are
still there. The media query is changed from 650 to 950 to support
half-screens.

For the nav-bar, a section was added and the indentation cleaned up.
Also height was adjusted so that the div expands when the text wraps in
and doesn't overlay the featured section.